ajustes varios en stadisticas
git-svn-id: http://proyecto-javaee.googlecode.com/svn/trunk@151 a03e1d51-c13a-4d2a-65c6-6c5e8594b909

+++ tms-api/src/main/java/org/inftel/tms/services/AffectedFacadeRemote.java
@@ -28,5 +28,5 @@ public interface AffectedFacadeRemote {
 
     int count();
 
-    int countByType(AffectedType type);
+    Long countByType(AffectedType type);
 }
