+++ tms-api/src/main/java/org/inftel/tms/services/AlertFacadeRemote.java
@@ -35,5 +35,5 @@ public interface AlertFacadeRemote {
 
     List<Alert> findAlertsByAffected(Person affected);
 
-    int countByType(AlertType type, Date fromDate, Date toDate);
+    Long countByType(AlertType type, Date fromDate, Date toDate);
 }
